backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-vertexai
langchain-community
psycopg2-binary
sqlalchemy
google-cloud-aiplatform
